backend/database.py
"""
MongoDB database connection and configuration.
"""

from motor.motor_asyncio import AsyncIOMotorClient
from typing import Optional
import os
from dotenv import load_dotenv

load_dotenv()

# MongoDB connection settings
MONGO_URL = os.getenv("MONGO_URL", "mongodb://localhost:27017")
MONGO_DB_NAME = os.getenv("MONGO_DB_NAME", "vibecation")

# Global database client
client: Optional[AsyncIOMotorClient] = None
database = None


async def connect_to_mongo():
    """Create database connection."""
    global client, database
    try:
        client = AsyncIOMotorClient(MONGO_URL)
        database = client[MONGO_DB_NAME]
        # Test the connection
        await client.admin.command('ping')
        print(f"✅ Connected to MongoDB: {MONGO_DB_NAME}")
    except Exception as e:
        print(f"❌ Failed to connect to MongoDB: {e}")
        raise


async def close_mongo_connection():
    """Close database connection."""
    global client
    if client:
        client.close()
        print("✅ MongoDB connection closed")


def get_database():
    """Get the database instance."""
    return database

